add search bar

+++ app/src/main/java/com/example/weatherforecast/model/WeatherDetails.kt
@@ -3,15 +3,15 @@ package com.example.weatherforecast.model
 data class WeatherDetails(
     val clouds: Int,
     val deg: Int,
-    val dt: Int,
+    val dt: Long,
     val feels_like: FeelsLike,
     val gust: Double,
     val humidity: Int,
     val pop: Double,
     val pressure: Int,
     val speed: Double,
-    val sunrise: Int,
-    val sunset: Int,
+    val sunrise: Long,
+    val sunset: Long,
     val temp: Temp,
     val weather: List<WeatherObject>
 )
